requirements.txt
Django==2.1.7
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-simple-cookie-consent==0.1.1
django-smtp-ssl==1.0
future==0.17.1
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.9
sendgrid==3.6.5
whitenoise==4.1.2
